Remove hot module replacement
